Add signature module
Because things other than transactions can be signed.
