+++ src/lib.rs
@@ -3,6 +3,7 @@ pub mod log;
 pub mod logger;
 pub mod event;
 pub mod transaction;
+pub mod signature;
 pub mod genesis;
 pub mod historian;
 pub mod accountant;
